Moving config to src
